implement models: Entity,Quote, IexQuote

+++ springboot/src/main/java/ca/jrvs/apps/trading/model/domain/IexQuote.java
@@ -1,11 +1,6 @@
 package ca.jrvs.apps.trading.model.domain;
 
-import java.util.HashMap;
-import java.util.Map;
 import javax.annotation.Generated;
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -153,11 +148,11 @@ public class IexQuote {
   @JsonProperty("avgTotalVolume")
   private Integer avgTotalVolume;
   @JsonProperty("iexBidPrice")
-  private Integer iexBidPrice;
+  private Double iexBidPrice;
   @JsonProperty("iexBidSize")
   private Integer iexBidSize;
   @JsonProperty("iexAskPrice")
-  private Integer iexAskPrice;
+  private Double iexAskPrice;
   @JsonProperty("iexAskSize")
   private Integer iexAskSize;
   @JsonProperty("iexOpen")
@@ -586,12 +581,12 @@ public class IexQuote {
   }
 
   @JsonProperty("iexBidPrice")
-  public Integer getIexBidPrice() {
+  public Double getIexBidPrice() {
     return iexBidPrice;
   }
 
   @JsonProperty("iexBidPrice")
-  public void setIexBidPrice(Integer iexBidPrice) {
+  public void setIexBidPrice(Double iexBidPrice) {
     this.iexBidPrice = iexBidPrice;
   }
 
@@ -606,12 +601,12 @@ public class IexQuote {
   }
 
   @JsonProperty("iexAskPrice")
-  public Integer getIexAskPrice() {
+  public Double getIexAskPrice() {
     return iexAskPrice;
   }
 
   @JsonProperty("iexAskPrice")
-  public void setIexAskPrice(Integer iexAskPrice) {
+  public void setIexAskPrice(Double iexAskPrice) {
     this.iexAskPrice = iexAskPrice;
   }
